requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses==0.6
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
idna==3.4
jsonschema==4.17.3
marshmallow==3.19.0
packaging==23.1
payme-pkg==2.2
Pillow==9.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uuid==1.30
